+++ Sources/Geometria/2D/Protocols/Vector/Vector2Type.swift
@@ -1,5 +1,5 @@
 /// Protocol for types that can represent 2D vectors.
-public protocol Vector2Type: VectorTakeable {
+public protocol Vector2Type: VectorTakeable where TakeDimensions == Vector2TakeDimensions {
     associatedtype SubVector2 = Self
 
     /// The X coordinate of this 2D vector.
@@ -14,6 +14,12 @@ public protocol Vector2Type: VectorTakeable {
     /// Initializes a new instance of this `Vector2Type` type by copying the
     /// coordinates of another `Vector2Type` of matching scalar type.
     init<Vector: Vector2Type>(_ vector: Vector) where Vector.Scalar == Scalar
+}
+
+/// Defines the dimension of an indexed takeable getter for a Vector 2 type.
+public enum Vector2TakeDimensions: Int {
+    case x
+    case y
 }
 
 public extension Vector2Type {